% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse.R
\name{browse.download_list}
\alias{browse.download_list}
\title{Open a browser window to display a Neotoma dataset within the Neotoma Explorer}
\usage{
\method{browse}{download_list}(x)
}
\arguments{
\item{x}{A \code{download_list} object.}
}
\description{
Using a \code{download} or \code{dataset} object, open up a browser window in the users default browser. Passing a \code{download_list} or \code{dataset_list} will open Neotoma Explorer with the first object and return a warning.
}

