\name{check_tables}
\alias{check_tables}
\title{Check to make sure that neotoma linked tables are up to date.}
\usage{
check_tables()
}
\value{
A warning, if the tables stored in the data folder are not
the same as the ones online.
}
\description{
A function to access the Neotoma API and return tables
corresponding to the parameters defined by the user.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org API
Reference: http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

