% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compile_downloads}
\alias{compile_downloads}
\title{Function to convert multiple downloads into a single large table.}
\usage{
compile_downloads(downloads)
}
\arguments{
\item{downloads}{A list of downloads as returned by \code{get_download}, or mutliple sites joined in a list.}
}
\value{
This command returns a \code{data.frame}

 To support further synthesis and analysis \code{compile_download} works to transform a list
 returned by \code{get_download} into a large \code{data.frame} with columns for site and sample attributes
 and also with the associated assemblage data at each sample depth.  This function also does the same for
 single sites.
}
\description{
From the assemblage data for multiple cores, return a single \code{data.frame} with columns for site
metadata and assemblage data.
}
\examples{
\dontrun{
#  Search for sites with "Thuja" pollen that are older than 8kyr BP and
#  that are on the west coast of North America:
t8kyr.datasets <- get_dataset(taxonname='Thuja*', loc=c(-150, 20, -100, 60), ageyoung = 8000)

#  Returns 3 records (as of 04/04/2013), get dataset for the first record, Gold Lake Bog.
thuja.sites <- get_download(sapply(t8kyr.datasets, function(x)x$DatasetID))

gold.p25 <- lapply(thuja.sites, compile_taxa, 'P25')

all.gold <- compile_downloads(gold.p25)

}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org

Gavin DG, Oswald WW, Wahl ER, Williams JW. 2003. A statistical approach to evaluating distance metrics and analog assignments for pollen records. Quaternary Research 60: 356-367.

Whitmore J, Gajewski K, Sawada M, Williams JW, Shuman B, Bartlein PJ, Minckley T, Viau AE, Webb III T, Shafer S, Anderson P, Brubaker L. 2005. Modern pollen data from North America and Greenland for multi-scale paleoenvironmental applications. Quaternary Science Reviews 24: 1828-1848.

Williams J, Shuman B. 2008. Obtaining accurate and precise environmental reconstructions from the modern analog technique and North American surface pollen dataset. Quaternary Science Reviews. 27:669-687.

API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

