% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\alias{counts.download}
\alias{counts.download_list}
\title{Access proxy count data}
\usage{
counts(obj, ...)

\method{counts}{download}(obj, ...)

\method{counts}{download_list}(obj, ...)
}
\arguments{
\item{obj}{an R object from which counts are to be extracted.}

\item{...}{arugments passed to other methods.}
}
\value{
Either a data frame of counts or a list of such objects.
}
\description{
Extract pollen or other proxy counts from data objects and returns them in a useful format.
}
\details{
Methods are available for "download" and "download_list" objects.
}
\author{
Gavin Simpson
}

