% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download}
\alias{download}
\title{A class for download objects.}
\description{
A \code{download} is an object with the full record for a single dataset.

A \code{download} is an object with the full record for a single dataset.
}
\details{
TO DO

TO DO
}
\author{
Simon Goring

Simon Goring
}
\keyword{methods}

