% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chroncontrol.R
\name{get_chroncontrol.dataset}
\alias{get_chroncontrol.dataset}
\title{Function to return chronological control tables from a \code{dataset}.}
\usage{
\method{get_chroncontrol}{dataset}(x, chronology = 1, verbose = TRUE,
  add = FALSE)
}
\arguments{
\item{x}{A \code{dataset}.}

\item{chronology}{When \code{download} objects have more than associated chronology, which chronology do you want?  Default is \code{1}.}

\item{verbose}{logical; should messages on API call be printed?}

\item{add}{Should the \code{chroncontrol} be added to the download object (only accepts \code{FALSE})}
}
\description{
Using a \code{dataset}, return the default chron-control table.
}
