% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chroncontrol.R
\name{get_chroncontrol.download}
\alias{get_chroncontrol.download}
\title{Function to return chronological control tables from a \code{download} object.}
\usage{
\method{get_chroncontrol}{download}(x, chronology = 1, verbose = TRUE,
  add = FALSE)
}
\arguments{
\item{x}{A single \code{download} object.}

\item{chronology}{For \code{download} methods, which chronology controls should be used?}

\item{verbose}{logical; should messages on API call be printed?}

\item{add}{Should the \code{chroncontrol} be added to the download object (default \code{FALSE})}
}
\description{
Using a \code{download}, return the default chron-control table as a \code{data.frame}.
}
