% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_chroncontrol.R
\name{get_chroncontrol.download_list}
\alias{get_chroncontrol.download_list}
\title{Function to return chronological control tables from a \code{download_list} object.}
\usage{
\method{get_chroncontrol}{download_list}(x, verbose = TRUE)
}
\arguments{
\item{x}{A \code{download_list} object.}

\item{verbose}{logical; should messages on API call be printed?}
}
\description{
Using a \code{download_list}, return the default chron-control table as a \code{data.frame}.
}

