% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_contacts}
\alias{get_contacts}
\title{Get contact information.}
\usage{
get_contacts(contactid, contactname, contactstatus, familyname)
}
\arguments{
\item{contactid}{<what param does>}

\item{contactname}{<what param does>}

\item{contactstatus}{<what param does>}

\item{familyname}{<what param does>}
}
\description{
Deprecated - see \code{get_contact}.
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

