% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset.download}
\alias{get_dataset.download}
\title{Obtain dataset information from an existing \code{download} object.}
\usage{
\method{get_dataset}{download}(x, ...)
}
\arguments{
\item{x}{An object of class \code{download}.}

\item{...}{objects passed from the generic.  Not used in the call.}
}
\description{
A function to access a \code{dataset} within a \code{download} object.
}
