% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset.geochronologic_list}
\alias{get_dataset.geochronologic_list}
\title{Obtain dataset information from an object of class \code{geochronologic_list}.}
\usage{
\method{get_dataset}{geochronologic_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{geochronologic_list}.}

\item{...}{objects passed from the generic.  Not used in the call.}
}
\description{
A function to access the Neotoma API and return datasets corresponding to the parameters defined by the user.
}
