% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset.numeric}
\alias{get_dataset.numeric}
\title{Obtain dataset information from a vector of dataset IDs.}
\usage{
\method{get_dataset}{numeric}(x = NULL, ...)
}
\arguments{
\item{x}{A single numeric dataset id, or a numeric vector.}

\item{...}{objects passed from the generic.  Not used in the call.}
}
\description{
A function to access the Neotoma API and return datasets corresponding to the parameters defined by the user.
}
