% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_dataset}
\alias{get_dataset}
\title{Obtain dataset information from the Neotoma Paleoecological Database.}
\usage{
get_dataset(x, ...)
}
\arguments{
\item{x}{<what param does>}

\item{...}{Optional additional arugments}
}
\value{
More details on the use of these parameters can be obtained from
   \url{http://api.neotomadb.org/doc/resources/datasets.}

   A list, with each item corresponding to an individual record.  Each list item
   (each dataset record) includes the following components:

 \item{ \code{DatasetID} }{Unique database record identifier for the dataset.}
 \item{ \code{DatasetName}  }{Name of the dataset; not commonly used.}
 \item{ \code{CollUnitHandle}  }{Code name of the Collection Unit with which the dataset is associated. This code may be up to 10 characters. Data are frequently distributed by Collection Unit, and the Handle is used for file names.}
 \item{ \code{CollUnitID}  }{Unique database record identifier for the collection unit.}
 \item{ \code{CollType}  }{The collection type. Types include cores, sections, excavations, and animal middens.}
 \item{ \code{DatasetType}  }{The dataset type, such as: geochronologic, loss-on-ignition, pollen, plant macrofossils, vertebrate fauna, etc.}
 \item{ \code{AgeOldest}  }{The oldest of all sample ages (in calendar years before present) in the dataset.}
 \item{ \code{AgeYoungest}  }{The youngest of all sample ages (in calendar years before present) in the dataset.}
 \item{ \code{SubDates}  }{An array of objects that describe dataset submission events.  If multiple submissions occured then this is a table.}
 \item{ \code{DatasetPIs}  }{An array of objects that describe Principal Investigators associated with a dataset.}
 \item{ \code{Site}  }{An object describing the site where the dataset samples were taken.}
}
\description{
A function to access the Neotoma API and return datasets corresponding to the parameters defined by the user.
}
\examples{
\dontrun{
# Search for sites with "Thuja" pollen that are older than 8kyr BP and
# that are on the west coast of North America:
t8kyr.datasets <- get_dataset(taxonname='Thuja*', loc=c(-150, 20, -100, 60), ageyoung = 8000)

# Search for vertebrate fossils in Canada (gpid: 756) within the last 2kyr.
gpids <- get_table(table.name='GeoPoliticalUnits')
canID <- gpids[which(gpids$GeoPoliticalName == 'Canada'),1]

v2kyr.datasets <- get_dataset(datasettype='vertebrate fauna', gpid=canID, ageold = 2000)
}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

