% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset.site}
\alias{get_dataset.site}
\title{Obtain dataset information from an existing \code{site} object.}
\usage{
\method{get_dataset}{site}(x, ...)
}
\arguments{
\item{x}{An object of class \code{site}.}

\item{...}{objects passed from the generic.  Not used in the call.}
}
\description{
A function to access the Neotoma API and return datasets corresponding to the parameters defined by the user.
}
