% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_datasets}
\alias{get_datasets}
\title{Obtain dataset information from the Neotoma Paleoecological Database.}
\usage{
get_datasets(siteid, datasettype, piid, altmin, altmax, loc, gpid, taxonids,
  taxonname, ageold, ageyoung, ageof, subdate)
}
\arguments{
\item{siteid}{<what param does>}

\item{datasettype}{<what param does>}

\item{piid}{<what param does>}

\item{altmin}{<what param does>}

\item{altmax}{<what param does>}

\item{loc}{<what param does>}

\item{gpid}{<what param does>}

\item{taxonids}{<what param does>}

\item{taxonname}{<what param does>}

\item{ageold}{<what param does>}

\item{ageyoung}{<what param does>}

\item{ageof}{<what param does>}

\item{subdate}{<what param does>}
}
\description{
Deprecated - see \code{get_dataset}.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

