% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_download.dataset}
\alias{get_download.dataset}
\title{<brief desc>}
\usage{
\method{get_download}{dataset}(dataset, verbose = TRUE)
}
\arguments{
\item{dataset}{<what param does>}

\item{verbose}{= TRUE <what param does>}
}
\description{
<full description>
}

