% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_download.R
\name{get_download.dataset_list}
\alias{get_download.dataset_list}
\title{Function to return full download records using a \code{dataset_list}.}
\usage{
\method{get_download}{dataset_list}(x, verbose = TRUE)
}
\arguments{
\item{x}{An object of class \code{dataset_list}.}

\item{verbose}{logical; should messages on API call be printed?}
}
\description{
Using a \code{dataset_list}, return all records associated with the data as a \code{download_list}.
}
