% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_download.default}
\alias{get_download.default}
\title{foo}
\usage{
\method{get_download}{default}(datasetid, verbose = TRUE)
}
\arguments{
\item{datasetid}{A single numeric dataset ID or a vector of numeric dataset IDs as returned by \code{get_datasets}.}

\item{verbose}{logical; should messages on API call be printed?}
}
\description{
foo
}

