% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publication.R
\name{get_publication.dataset}
\alias{get_publication.dataset}
\title{A function to get publications for datasets in the Neotoma Database using the API.}
\usage{
\method{get_publication}{dataset}(x, ...)
}
\arguments{
\item{x}{an object of class \code{dataset}.}

\item{...}{objects passed from the generic.  Not used in the call.}
}
\description{
The function takes a \code{dataset} and returns a table with publication information from the Neotoma Paleoecological Database.
}
