% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publication.R
\name{get_publication.download}
\alias{get_publication.download}
\title{A function to get publications for downloads in the Neotoma Database using the API.}
\usage{
\method{get_publication}{download}(x, ...)
}
\arguments{
\item{x}{an object of class \code{download}.}

\item{...}{objects passed from the generic.  Not used in the call.}
}
\description{
The function takes a \code{download} and returns a table with publication information from the Neotoma Paleoecological Database.
}
