% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_site.default}
\alias{get_site.default}
\title{title}
\usage{
\method{get_site}{default}(sitename, altmin, altmax, loc, gpid,
  download = NULL)
}
\arguments{
\item{sitename}{A character string representing the full or partial site name.}

\item{altmin}{Minimum site altitude  (in m).}

\item{altmax}{Maximum site altitude (in m).}

\item{loc}{A numeric vector c(lonW, latS, lonE, latN) representing the bounding box within which to search for sites.  The convention here is to use negative values for longitudes west of Grewnwich or longitudes south of the equator.}

\item{gpid}{A character string or numeric value, must correspond to a}

\item{download}{= NULL <what param does>}
}
\description{
title
}

