% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_site.R
\name{get_site.geochronologic}
\alias{get_site.geochronologic}
\title{Return Site Information from a \code{geochronologic}}
\usage{
\method{get_site}{geochronologic}(x, ...)
}
\arguments{
\item{x}{An object of class \code{geochronologic}.}

\item{...}{Arguments passed from the generic method, not used.}
}
\description{
Return site information from the Neotoma Paleoecological Database.
}

