% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site.R
\name{get_site.numeric}
\alias{get_site.numeric}
\title{Return Site information from a vector of numeric elements.}
\usage{
\method{get_site}{numeric}(sitename = NULL, ...)
}
\arguments{
\item{sitename}{A numeric value or vector of numeric elements.}

\item{...}{Arguments passed from the generic method, not used.}
}
\description{
Return site information from the Neotoma Paleoecological Database.
}
