% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_site}
\alias{get_site}
\title{Return Site Information.}
\usage{
get_site(sitename, altmin, altmax, loc, gpid)
}
\arguments{
\item{sitename}{A character string representing the full or partial site name.}

\item{altmin}{Minimum site altitude  (in m).}

\item{altmax}{Maximum site altitude (in m).}

\item{loc}{A numeric vector c(lonW, latS, lonE, latN) representing the bounding box within which to search for sites.  The convention here is to use negative values for longitudes west of Grewnwich or longitudes south of the equator.}

\item{gpid}{A character string or numeric value, must correspond to a valid geopolitical identity in the Neotoma Database.  Use get.tables('GeoPoliticalUnits') for a list of acceptable values, or link here: http://api.neotomadb.org/apdx/geopol.htm}
}
\value{
A table:

\describe{
 \item{siteid}{Unique database record identifier for the site.}
 \item{sitename}{Name of the site.}
 \item{long}{Mean longitude, in decimal degrees, for a site (-180 to 180).}
 \item{lat}{Mean latitude, in decimal degrees, for a site (-90 to 90).}
 \item{elev}{Elevation in meters.}
 \item{description}{Free form description of a site, including such information as physiography and vegetation around the site.}
 \item{long_acc}{If the site is described by a bounding box this is the box width.}
 \item{lat_acc}{If the site is described by a bounding box this is the box height.}
}
}
\description{
Return site information from the Neotoma Paleoecological Database.

\code{get_site} returns site information from the Neotoma Paleoecological Database
   based on parameters defined by the user.
}
\examples{
\dontrun{
#  What is the distribution of site elevations in Neotoma?
all.sites <- get_site()  #takes a bit of time.

plot(density(all.sites$elev, from = 0, na.rm=TRUE),
main = 'Altitudinal Distribution of Neotoma Sites', xlab = 'Altitude (m)', log='x')
}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/sites
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

