% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_sites}
\alias{get_sites}
\title{Return Site Information.}
\usage{
get_sites(siteid, sitename, altmin, altmax, loc, gpid)
}
\arguments{
\item{siteid}{<what param does>}

\item{sitename}{<what param does>}

\item{altmin}{<what param does>}

\item{altmax}{<what param does>}

\item{loc}{<what param does>}

\item{gpid}{<what param does>}
}
\description{
Deprecated, see \code{get_site}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

