% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{gp.table}
\alias{gp.table}
\title{A list of all the geopolitical entities in the Neotoma database.}
\format{a \code{data.frame} object}
\source{
The Neotoma database.
}
\usage{
gp.table
}
\description{
A list of geopolitical entities with associated numeric ID values.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}

