% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{neotoma}
\alias{neotoma}
\alias{neotoma-package}
\title{neotoma}
\description{
neotoma
}
\details{
The neotoma package is a programmatic R interface to the \href{http://www.neotomadb.org/}{Neotoma Paleoecological Database}. The package is intended to both allow users to search for sites and to download data for use in analyical workflows of paleoecological research.
}

