% Generated by roxygen2 (4.0.1): do not edit by hand
\name{param_check}
\alias{param_check}
\title{Internal function to check passed parameters.}
\usage{
param_check(cl)
}
\arguments{
\item{cl}{Contact ID is a numerical value associated with the Neotoma
Contact table's numerical Contact ID.}
}
\value{
A list with two components:

\itemize{
 \item{flag}{Returns a 0 if everything's fine, a 1 if there's a problem.}
 \item{message}{A list of error messages.}
}
}
\description{
Functions \code{get_site}, \code{get_dataset} and others pass parameters to \code{param_check}, \code{param_check} tells them if there's a problem.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

