% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leaflet.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Leaflet plots for neotoma data.}
\usage{
plot_leaflet(x, providerTiles = "Stamen.TerrainBackground", ...)
}
\arguments{
\item{x}{A neotoma data object}

\item{providerTiles}{Default "Stamen.TerrainBackground", a character string indicating the tile background to be used for plotting.}

\item{...}{Other terms to be passed to the function.}
}
\value{
A \code{leaflet} object
}
\description{
A plotting function to provide interactive data investigation using the leaflet tools.
  This package requires a connection to the internet for proper functioning.
}
