\docType{data}
\name{pollen.equiv}
\alias{pollen.equiv}
\title{A table to convert the pollen taxa identified by investigators to standardized lists.}
\format{a \code{data.frame} object}
\description{
  A list of standardized (published) taxonomies from the
  literature to help standardize taxonomies for synthesis
  work.
}
\details{
  Taxon conversion table (readable).
}
\author{
  Simon J. Goring \email{simon.j.goring@gmail.com};
  Jeremiah Marsicek
}
\keyword{datasets}

