% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.download}
\alias{print.download}
\title{Print function for the `download` class.}
\usage{
\method{print}{download}(x)
}
\value{
Returns a very brief discription of the size of the `download` object.
}
\description{
Print function for the `download` class.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

