% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.site}
\alias{print.site}
\title{Print function for the \code{site} class.}
\usage{
\method{print}{site}(x)
}
\value{
Returns a very brief discription of the \code{site} object as a \code{data.frame}
}
\description{
Print function for the \code{site} class.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

