% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_methods.R
\name{read.tilia}
\alias{read.tilia}
\title{Read proxy data from Tilia TLX files}
\usage{
read.tilia(file)
}
\arguments{
\item{file}{a string representing a Tilia TLX format file.}
}
\value{
Return a `download` object.
}
\description{
Read proxy data from a Tilia TLX format file.
}
\examples{
\dontrun{
  crystal <- read.tilia('crystal.tlx')
}

}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
