\docType{data}
\name{taxon.list}
\alias{taxon.list}
\title{Neotoma taxon list}
\format{a \code{data.frame} object}
\source{
  The Neotoma database.
}
\description{
  The taxonomy table for datasets in neotoma, as would be
  returned by \code{get_table}
}
\author{
  Simon J. Goring \email{simon.j.goring@gmail.com}
}
\keyword{datasets}

