% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{translate.table}
\alias{translate.table}
\title{A table to convert the original taxa to standardized lists.}
\format{a \code{data.frame} object}
\source{
The Neotoma database.
}
\usage{
translate.table
}
\description{
A list of standardized (published) taxonomies from the literature to help standardize taxonomies for synthesis work.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}

