% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse.download_list}
\alias{browse.download_list}
\title{Open a browser window to display a Neotoma dataset within the Neotoma Explorer}
\usage{
\method{browse}{download_list}(x)
}
\arguments{
\item{x}{A \code{download_list} object.}
}
\description{
Using a numeric value, \code{download}, \code{download_list}, \code{dataset} or \code{dataset_list} object, open up a browser window in the users default browser. Very large objects
}
