% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{download}
\alias{download}
\alias{download-package}
\title{A class for download objects.}
\description{
A \code{download} is an object with the full record for a single dataset.
}
\details{
TO DO
}
\author{
Simon Goring
}
\keyword{methods}

