% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamodel.R, R/download.R
\name{download}
\alias{download}
\alias{download}
\title{A class for download objects.}
\description{
A \code{download} is an object with the full record for a single dataset.

A \code{download} is an object with the full record for a single dataset.
}
\details{
TO DO

TO DO
}
\author{
Simon Goring

Simon Goring
}
\keyword{methods}
