% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chroncontrol.R
\name{get_chroncontrol.default}
\alias{get_chroncontrol.default}
\title{Function to return chronological control tables from a chronologic ID.}
\usage{
\method{get_chroncontrol}{default}(x, chronology = 1, verbose = TRUE,
  add = FALSE)
}
\arguments{
\item{x}{A single numeric chronology ID or a vector of numeric chronology IDs as returned by \code{get_datasets}.}

\item{chronology}{For \code{download} methods, which chronology controls should be used?}

\item{verbose}{logical; should messages on API call be printed?}

\item{add}{logical, should this chron control be added to the download object?}
}
\description{
Using the chronology ID, return the chron control table as a \code{data.frame}.
}
