% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chroncontrol.R
\name{get_chroncontrol.download_list}
\alias{get_chroncontrol.download_list}
\title{Function to return chronological control tables from a \code{download_list} object.}
\usage{
\method{get_chroncontrol}{download_list}(x, chronology = 1,
  verbose = TRUE, add = FALSE)
}
\arguments{
\item{x}{A \code{download_list} object.}

\item{chronology}{When \code{download} objects have more than associated chronology, which chronology do you want?  Default is \code{1}.}

\item{verbose}{logical; should messages on API call be printed?}

\item{add}{Should the \code{chroncontrol} be added to the download object (default \code{FALSE})}
}
\description{
Using a \code{download_list}, return the default chron-control table as a \code{data.frame}.
}
