% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_contacts}
\alias{get_contacts}
\title{Get contact information.}
\usage{
get_contacts(contactid, contactname, contactstatus, familyname)
}
\arguments{
\item{contactid}{Contact ID is a numerical value associated with the Neotoma
Contact table's numerical Contact ID.}

\item{contactname}{A character string indicating the data contributors' project,
organization or personal name.  May be a partial string and can include wildcards.}

\item{contactstatus}{The current status of the contact.  Possible values include:
active, deceased, defunct, extant, inactive, retired, unknown.}

\item{familyname}{A character string.  Full or partial string indicating the
   contact's last name.}
}
\description{
Deprecated - see \code{get_contact}.
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{IO}
\keyword{connection}

