% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_contacts}
\alias{get_contacts}
\title{Get contact information.}
\usage{
get_contacts(contactid, contactname, contactstatus, familyname)
}
\description{
Deprecated - see \code{get_contact}.
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{IO}
\keyword{connection}

