% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_datasets}
\alias{get_datasets}
\title{Obtain dataset information from the Neotoma Paleoecological Database.}
\usage{
get_datasets(siteid, datasettype, piid, altmin, altmax, loc, gpid, taxonids,
  taxonname, ageold, ageyoung, ageof, subdate)
}
\description{
Deprecated - see \code{get_dataset}.
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{API}
\keyword{Neotoma}
\keyword{Palaeoecology}

