% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site.R
\name{get_site.dataset_list}
\alias{get_site.dataset_list}
\title{Return Site Information from a \code{dataset_list}}
\usage{
\method{get_site}{dataset_list}(sitename, ...)
}
\arguments{
\item{sitename}{An object of class \code{dataset_list}.}

\item{...}{Arguments passed from the generic method, not used.}
}
\description{
Return site information from the Neotoma Paleoecological Database.
}
