% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Return Site Information.}
\usage{
get_sites(sitename, altmin, altmax, loc, gpid)
}
\arguments{
\item{sitename}{A character string representing the full or partial site name.}

\item{altmin}{Minimum site altitude  (in m).}

\item{altmax}{Maximum site altitude (in m).}

\item{loc}{A numeric vector c(lonW, latS, lonE, latN) representing the bounding box within which to search for sites.  The convention here is to use negative values for longitudes west of Grewnwich or longitudes south of the equator.}

\item{gpid}{A character string or numeric value, must correspond to a valid geopolitical identity in the Neotoma Database.  Use get.tables('GeoPoliticalUnits') for a list of acceptable values, or link here: http://api.neotomadb.org/apdx/geopol.htm}
}
\description{
Deprecated, see \code{get_site}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\keyword{IO}
\keyword{connection}

