% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_check.R
\name{param_check}
\alias{param_check}
\title{Internal function to check passed parameters.}
\usage{
param_check(cl)
}
\arguments{
\item{cl}{Contact ID is a numerical value associated with the Neotoma
Contact table's numerical Contact ID.}
}
\value{
A list with two components:

 \item{flag}{Returns a 0 if everything's fine, a 1 if there's a problem.}
 \item{message}{A list of error messages.}
}
\description{
Functions \code{\link{get_site}}, \code{\link{get_dataset}} and others pass parameters to \code{param_check}, which tells them if there's a problem.
}
\references{
Neotoma Project Website: http://www.neotomadb.org
API Reference:  http://api.neotomadb.org/doc/resources/contacts
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
\keyword{internal}
\keyword{misc}
