% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.table.R
\docType{data}
\name{taxon.list}
\alias{taxon.list}
\title{Neotoma taxon list}
\format{a \code{data.frame} object}
\source{
The Neotoma database.
}
\usage{
taxon.list
}
\description{
The taxonomy table for datasets in neotoma, as would be returned by \code{\link{get_table}}
}
\author{
Simon J. Goring \email{simon.j.goring@gmail.com}
}
